Added possibility to change some input bindings.
Now it is about some part of player control only: moving, shooting, reloading. Also now it is not enough convenient: there is no special screen to change bindings in-game.

Bindings are stored in config file. User can select a device (Type of binding) which can be 0 for keyboard and 1 for mouse. User can also select a value (Value of binding). For mouse the value can be 1 (left mouse button), 2 (middle mouse button), 3 (right mouse button) and so on. For keyboard the value should be the key SDL code which is very similar to ASCII code. See SDL_keysym.h for details ;) or wait for changes.